refactor(console): remove api-v4 entities
keep plan-v4 since rest-api-v2 not support plan creation yet
