+++ gravitee-apim-console-webui/src/entities/plan-v4/index.ts
@@ -14,3 +14,4 @@
  * limitations under the License.
  */
 export * from './plan';
+export * from './Flow';
