+++ gravitee-apim-console-webui/src/management/api/component/plan/2-secure-step/plan-edit-secure-step.component.ts
@@ -28,7 +28,7 @@ import { PolicyService } from '../../../../../services-ngx/policy.service';
 import { ResourceService } from '../../../../../services-ngx/resource.service';
 import { ResourceListItem } from '../../../../../entities/resource/resourceListItem';
 import { Api as ApiV3 } from '../../../../../entities/api';
-import { ApiEntity as ApiV4 } from '../../../../../entities/api-v4';
+import { ApiV4 } from '../../../../../entities/management-api-v2';
 import { SnackBarService } from '../../../../../services-ngx/snack-bar.service';
 
 const allSecurityTypes = [
